Move nop instance into InstanceData
With this we keep all instances in the same map and can treat the nop
instance (for the cri sandbox "container", aka "pause") the same as the
real implementation of `Instance`.

I did look into making `InstanceData` more polymorphic, but ran into
problems with the compiler complaining that it doesn't know the size at
compile time and that it can't create the trait object... so instead we
store the "base" on a separate field in the same `InstanceData<T>`, and
wrap them both in an Option type to tell which one is set.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>