Add unit tests for shim::Local
This neccessarily moves the details of the implementation to functions
on the Local struct rather because it is non-trivial to create a
TtrpcContext for test cases.

Instead we just have the implementation without TtrpcContext.
We may need to add timeout support (especially for wait()) which should
use the timeouts specified from TtrpcContext.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>