benchmarks: add initial benchmarking support. (#126)
Add a submodule "webassembly-benchmarks". The benchmark contains a set
of wasm files from libsodium. Only the benchmarks which are fast to run
(take under 5 seconds to complete on a desktop computer using WasmEdge)
were selected for benchmarking.

The benchmark suite initializes the Wasm engine, runs the benchmark, and
cleans up the container. This is done for both WasmEdge and Wasmtime
shim crates. Note that the results can't be directly compared right now,
because Wasmtime shim hasn't been converted to youki/libcontainer-rs yet
(and that changes the performance due to the extended namespacing).

The benchmarking library used is Criterion. Run the benchmarks like
this:

  # cargo bench

If you have gnuplot binary installed, Criterion will generate an HTML
report in target/criterion/report/index.html file.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>